+++ Ejemplo3.java
@@ -1,9 +1,9 @@
 public class Ejemplo3 {
 	public static void main(String[] args){
 
-		double squareArea = 5;
+		int squareArea = 5;
 
-		double area = (squareArea * squareArea);
+		int area = (squareArea * squareArea);
 
 		System.out.println("El area del cuadrado es: " + area);
 	}
