+++ Ejemplo4.java
@@ -1,14 +1,14 @@
 import java.util.Scanner;
 public class Ejemplo4 {
     public static void main (String argv[]) {
-        float side;
-        float squareArea;
+        int side;
+        int squareArea;
 
         System.out.println("Enter the side:");
 
         Scanner inputValue;
         inputValue = new Scanner(System.in);
-        side = inputValue.nextFloat();
+        side = inputValue.nextInt();
 
         squareArea = side * side;
         System.out.println(squareArea);
